fix(keys): auto-clear conflict banner, split oversized test file
Review follow-ups on the in-app keybinding conflict warning:

- Split the new KeybindingConflicts test suite out of
  KeybindingTests.swift (was 361 lines, over the 350-line hard
  ceiling in AGENTS.md §2.1) into its own
  KeybindingConflictsTests.swift; the original file is back to
  297 lines.

- Auto-clear the banner once the conflict it warned about is
  resolved, instead of only ever being dismissed manually:
  noteRecordedCombo warns if the just-recorded row conflicts,
  else clears the banner once the whole config has no conflict
  left, else leaves it as-is (some other row still conflicts,
  so the warning stays accurate). warnIfAnyConflict (Adopt,
  Lua-editor save) does a plain set-or-clear from the
  whole-config check, since those are explicit batch checks
  with no single row to compare against. The dismiss (x) button
  still works independently of this.

- Reworded the banner message to be self-contained ("see the
  ⚠️ mark below" instead of naming the tab it's already shown
  in), and made docs/configuration.md consistently say
  "Shortcuts" (its actual tab label) instead of mixing in
  "Keybindings tab".

- Corrected stale doc comments left over from the removed
  system-notification / app-launch design in
  KeybindingConflicts.swift (both doc comments) and
  SettingsModel.swift (keybindingWarning, save(), adoptIntoGui(),
  warnIfAnyConflict()) to describe the current banner behavior.

Fixes #10
